python/requirements.txt
openai
python-dotenv
requests 
Flask
Flask-CORS